No data screen added

+++ src/firebase/firebasedb/GetFromDb.js
@@ -4,6 +4,7 @@ import './getFromDb.css';
 import Context from '../firebasedb/validationContext/validationContext';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTrashAlt, faPencilAlt} from '@fortawesome/free-solid-svg-icons';
+import HomePageWithoutData from './validationContext/NoDataComponent';
 
 
 class GetFromDb extends React.Component {
@@ -39,33 +40,37 @@ class GetFromDb extends React.Component {
     }
     render() {
         const {data, error} = this.state;
-        return(
-            <div>
-                <ul className="getFromDb--list">
-                    {data.map((item, i) => (
-                        <li key={i}
-                            className="getFromDb--item__wrapper"
+        if (data.length <= 0 ) {
+            return <HomePageWithoutData/>
+        } else {
+            return (
+                <div>
+                    <ul className="getFromDb--list">
+                        {data.map((item, i) => (
+                            <li key={i}
+                                className="getFromDb--item__wrapper"
                             >
-                            <div className="getFromDb--item">
-                                <div className="getFromDb--item__category"> {item.category} </div>
-                                <div className="getFromDb--item__price"> {item.price}zł </div>
-                                {item.id &&
-                                <a onClick={() =>
-                                    db.collection('receits')
-                                        .doc(item.id)
-                                        .delete()
-                                        .then(this.props.toggle())
-                                }><FontAwesomeIcon icon={faTrashAlt} /></a>}
-                            </div>
-                           <div className="getFromDb--item__note">
-                                <FontAwesomeIcon icon={faPencilAlt}/> {item.note}
+                                <div className="getFromDb--item">
+                                    <div className="getFromDb--item__category"> {item.category} </div>
+                                    <div className="getFromDb--item__price"> {item.price}zł</div>
+                                    {item.id &&
+                                    <a onClick={() =>
+                                        db.collection('receits')
+                                            .doc(item.id)
+                                            .delete()
+                                            .then(this.props.toggle())
+                                    }><FontAwesomeIcon icon={faTrashAlt}/></a>}
                                 </div>
-                        </li>
-                    ))}
-                </ul>
-                {error && <p>{error.message}</p>}
-            </div>
-        )
+                                <div className="getFromDb--item__note">
+                                    <FontAwesomeIcon icon={faPencilAlt}/> {item.note}
+                                </div>
+                            </li>
+                        ))}
+                    </ul>
+                    {error && <p>{error.message}</p>}
+                </div>
+            )
+        }
     }
 }
